refactor and move to hooks folder
